requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
chromadb==0.4.22
pydantic==2.7.0
python-dotenv==1.1.0
numpy<2.0.0
sentence-transformers==2.3.1
ctranslate2==3.24.0
transformers==4.36.2
torch==2.1.2
requests==2.31.0
pymilvus==2.3.6 